factor out the supabase object obtaining code in a utils file
