+++ infra/lambdas/events/createEvent/index.js
@@ -3,14 +3,11 @@ const {
   DynamoDBDocumentClient, UpdateCommand,
 } = require("@aws-sdk/lib-dynamodb");
 
-const { createClient } = require("@supabase/supabase-js");
-const ENDPOINT = 'https://gsriiicvvxzvidaakctw.supabase.co';
-const EVENTS_TABLE = "Events";
+const { getSupabase } = require("../../shared_utils/supabase");
 
-const { SecretsManagerClient, GetSecretValueCommand } = require("@aws-sdk/client-secrets-manager");
-const REGION = process.env.AWS_REGION;
 const SUPABASE_SECRET_ID = process.env.SUPABASE_SECRET_ID;
-const secrets_client = new SecretsManagerClient({ region: REGION });
+const EVENTS_TABLE = "Events";
+const REGION = process.env.AWS_REGION;
 const ddb = DynamoDBDocumentClient.from(new DynamoDBClient({ region: REGION }));
 
 function dummyCognito(){
@@ -19,15 +16,6 @@ function dummyCognito(){
 
 function isAdmin(clientEmail){
     return dummyCognito()[0] === clientEmail;
-}
-
-async function getSupabase(){
-    const r = await secrets_client.send(new GetSecretValueCommand({ SecretId: SUPABASE_SECRET_ID }));
-    const raw = r.SecretString ?? Buffer.from(r.SecretBinary || "", "base64").toString("utf8");
-    const obj = JSON.parse(raw); 
-    const api_key = obj.SUPABASE_SECRET_KEY;
-    const cachedSupabase = createClient(ENDPOINT, api_key);
-    return cachedSupabase;
 }
 
 exports.handler = async (event) => {
@@ -58,7 +46,7 @@ exports.handler = async (event) => {
         const eventDeadline = parameters.event_deadline;
         const eventLocation = parameters.event_location;
 
-        const supabase = await getSupabase();
+        const supabase = await getSupabase(SUPABASE_SECRET_ID, REGION);
 
         const response = await supabase.from(EVENTS_TABLE).insert({
             event_id: newEventId,
